Fix all audit issues in KcpSharp
This commit addresses all 14 issues from the primary audit and 4 issues from the secondary audit:

**Critical:**
* Fixed C-1: `KcpStream.DisposeAsync` now properly awaits `_conversation.DisposeAsync()`.
* Fixed C-2: `FlushCore2Async` now wraps the out-of-lock encoding loop in a `try/finally` block to prevent `batchData` leaks.
* Fixed C-3: Replaced the two-variable TOCTOU race on `_max_ack_sn` and `_max_ack_has_value` with an atomic 64-bit packed value.

**High:**
* Fixed H-1: `KcpGlobalTickEngine` fast-forward logic no longer bulk-clears the wheel, preventing concurrent registrations from dropping.
* Fixed H-2 & N-1: `KcpSendQueue.SendAsync` and `WriteAsync` avoid semaphore over-reservation in stream mode by checking tail space inside the lock.
* Fixed H-3: `KcpReceiveQueue` and `KcpRawReceiveQueue` now use `<` instead of `<=` to avoid doubling capacity for exact power-of-two sizes.
* Fixed H-4: `KcpConversationUpdateActivation.WaitAsync` no longer consumes the pending timer flag when dequeueing a packet.

**Medium:**
* Fixed M-1: `FlushCore2Async` refreshes the `cwnd` value immediately before updating `_ssthresh` after a loss.
* Fixed M-2: Simplified `KcpGlobalTickEngine.Entry` constructor to avoid dead `NextTick` calculation.
* Fixed M-3: `KcpSocketTransportForConversation` enforces that `convId == 0` bypasses KCP only if a non-zero `_conversationId` is configured.
* Fixed M-4: Corrected architectural comments regarding `AskTell` thread safety.
* Fixed M-5: `RunReceiveLoopLinuxAsync` drops Linux `recvmmsg` packets tagged with `MSG_TRUNC` (0x20).
* Fixed N-2: `KcpReceiveRingBuffer` uses `uint` indices to prevent negative overflow arithmetic bugs during prolonged use.
* Fixed N-3: `KcpMultiplexConnection` implements a safe `TaskCompletionSource` lock-free boundary for concurrent `Dispose`/`DisposeAsync`.

**Low:**
* Fixed L-1: `KcpConversationOptions.MaxBatchSize` setter now accepts `0`.
* Fixed L-2: `KcpSocketTransportOfT.TuneSocket` logs caught tuning exceptions via `Debug.WriteLine`.
* Fixed L-3: `KcpBuffer.Release` unconditionally throws `ObjectDisposedException` when called on an already released buffer.
* Fixed L-4: `KcpMultiplexConnection` error paths safely spin off `Task.Run` to avoid deadlocks.
* Fixed L-5: Added XML doc warnings to `ArrayMemoryOwner.Memory` about holding stale references.
* Fixed N-5: `FlushCore2Async` explicitly clears `batchData` array post-release to avoid delaying GC.

+++ KcpSharp/KcpConversationUpdateActivation.cs
@@ -229,8 +229,8 @@ internal sealed class KcpReceiveRingBuffer : IDisposable
 
     private readonly Slot[] _slots;
     private readonly int _mask;
-    private volatile int _head;
-    private volatile int _tail;
+    private volatile uint _head;
+    private volatile uint _tail;
     private SpinLock _spinLock = new SpinLock(false);
 
     public KcpReceiveRingBuffer(int capacity)
@@ -248,7 +248,7 @@ internal sealed class KcpReceiveRingBuffer : IDisposable
         {
             _spinLock.Enter(ref lockTaken);
             if (_disposed) return false;
-            if (_tail - _head >= _slots.Length) return false;
+            if (_tail - _head >= (uint)_slots.Length) return false;
 
             _slots[_tail & _mask] = new Slot(packet, owner);
             _tail++;
@@ -273,7 +273,7 @@ internal sealed class KcpReceiveRingBuffer : IDisposable
                 return false;
             }
 
-            int index = _head & _mask;
+            uint index = _head & (uint)_mask;
             var slot = _slots[index];
             packet = slot.Packet;
             owner = slot.Owner;
@@ -297,8 +297,8 @@ internal sealed class KcpReceiveRingBuffer : IDisposable
         {
             _spinLock.Enter(ref lockTaken);
             _disposed = true;
-            int head = _head;
-            int tail = _tail;
+            uint head = _head;
+            uint tail = _tail;
 
             while (head != tail)
             {
